pyproject.toml
[tool.poetry]
name = "obsidian-review-notes"
version = "0.1.0"
description = ""
authors = ["Tamir Bahar"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
jinja2 = "^3.1.2"
typer = "^0.7.0"
attrs = "^23.1.0"
gitpython = "^3.1.31"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.263"
isort = "^5.12.0"
nox = "^2023.4.22"
mypy = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
review = 'obsidian_review_notes.main:main'

[tool.isort]
line_length = 88

[tool.black]
line-length = 88